Update 3rd exercise
